requirements.txt
plaid-python>=14.0.0
Pillow>=10.0.0
requests>=2.28.0
flask>=3.0.0
